chore: remove unused cutscene code and clarify project history
Removing CutsceneScreen.java as it is not required for the current MVP.

NOTE ON HISTORY: The previous large commit (23.12.2025) (adding ~2000 lines) was an accidental bulk push that included the core 'map.properties' data and the physics engine setup. The high line count is primarily due to the map data files, not generated code. The map included 1400 lines of data!

Within this Bulky commit I also did the following:
- Made the Debris
- Made the Destructible Interface
- Exit and SHOVEL both as HiddenObject

=> Implemented those in the Player.java Class and the GameMap.java